Add utils for getting roles and username/sub from requests
